rename class controllers
