+++ src/controllers/auth/sendPublicProfile.ts
