+++ src/controllers/auth/signout.ts
