+++ src/controllers/auth/updateAvatar.ts
