+++ src/controllers/auth/updatePassword.ts
