+++ src/controllers/product/deleteProductImage.ts
